+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Add
 - Add test option for DictZip CLI.
+- Add DictZipInputStream consructor that get a filename as argument.
 
 ### Changed
 - README: there is no test dependency for commons-io.
+- Default buffer size in the DictZipInputStream was changed to 4096 byte.
 
 ### Fixed
 - Check return value of InputStream.skip() method in the utility method